backend/requirements.txt
numpy
scikit-learn
transformers
torch
nltk
fuzzywuzzy
pandas
tqdm
requests
beautifulsoup4
PyPDF2
scholarly
python-dotenv
serpapi==0.1.5
fastapi
uvicorn
python-multipart
pydantic
aiofiles 